Split converter package into files.
