+++ internal/converter/consts.go
@@ -1,0 +1,10 @@
+package converter
+
+const (
+	metadataEncoding          = "encoding"
+	metadataEncodingBinary    = "binary/plain"
+	metadataEncodingJSON      = "json/plain"
+	metadataEncodingNil       = "binary/null"
+	metadataEncodingProtoJSON = "json/protobuf"
+	metadataEncodingProto     = "binary/protobuf"
+)
