+++ internal/converter/value.go
@@ -1,0 +1,19 @@
+package converter
+
+type (
+	// Value is used to encapsulate/extract encoded value from workflow/activity.
+	Value interface {
+		// HasValue return whether there is value encoded.
+		HasValue() bool
+		// Get extract the encoded value into strong typed value pointer.
+		Get(valuePtr interface{}) error
+	}
+
+	// Values is used to encapsulate/extract encoded one or more values from workflow/activity.
+	Values interface {
+		// HasValues return whether there are values encoded.
+		HasValues() bool
+		// Get extract the encoded values into strong typed value pointers.
+		Get(valuePtr ...interface{}) error
+	}
+)
